refactor: separate ProductEditAccordion and Form component
